requirements.txt
duffel_api
requests
python-dotenv